# openapi/lis-internal-v1.yaml
openapi: 3.1.0
info:
  title: LIS Core API
  summary: Starter modular monolith for a laboratory information system
  description: Internal API with working persistence for auth, master data, orders,
    specimen lifecycle, task orchestration, observations, reports, audit, provenance,
    device registry, HL7 v2 integrations, device gateway, autoverification, ASTM-style
    drivers, and a read/search FHIR facade.
  version: 0.7.0
paths:
...
  /fhir/R4/Task:
    get:
      tags:
      - fhir
      summary: Search Tasks
      operationId: search_tasks_fhir_R4_Task_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: focus
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Focus
      - name: code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Patient
      - name: _id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: ' Id'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Tasks Fhir R4 Task Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fhir/R4/Task/{task_id}:
    get:
      tags:
      - fhir
      summary: Read Task
      operationId: read_task_fhir_R4_Task__task_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Task Fhir R4 Task  Task Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/tasks:
    post:
      tags:
      - tasks
      summary: Create Task
      operationId: create_task_api_v1_tasks_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - tasks
      summary: List Tasks
      operationId: list_tasks_api_v1_tasks_get
      security:
      - HTTPBearer: []
      parameters:
      - name: queue
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Queue
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: owner_user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/TaskSummary'
                title: Response List Tasks Api V1 Tasks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}/claim:
    post:
      tags:
      - tasks
      summary: Claim Task
      operationId: claim_task_api_v1_tasks__task_id__claim_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskActionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}/start:
    post:
      tags:
      - tasks
      summary: Start Task
      operationId: start_task_api_v1_tasks__task_id__start_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/TaskActionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}/pause:
    post:
      tags:
      - tasks
      summary: Pause Task
      operationId: pause_task_api_v1_tasks__task_id__pause_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskActionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}/complete:
    post:
      tags:
      - tasks
      summary: Complete Task
      operationId: complete_task_api_v1_tasks__task_id__complete_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/TaskActionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}/fail:
    post:
      tags:
      - tasks
      summary: Fail Task
      operationId: fail_task_api_v1_tasks__task_id__fail_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskActionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateTaskRequest:
      properties:
        group_identifier:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Identifier
        based_on_order_item_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Based On Order Item Id
        focus_type:
          $ref: '#/components/schemas/TaskFocusType'
        focus_id:
          type: string
          format: uuid
          title: Focus Id
        queue_code:
          type: string
          title: Queue Code
        status:
          type: string
          title: Status
        business_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Business Status
        priority:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority
        owner_user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner User Id
        owner_practitioner_role_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Practitioner Role Id
        device_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Device Id
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
        inputs:
          additionalProperties: true
          type: object
          title: Inputs
      additionalProperties: false
      type: object
      required:
      - focus_type
      - focus_id
      - queue_code
      - status
      title: CreateTaskRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TaskActionRequest:
      properties:
        owner_user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner User Id
        owner_practitioner_role_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Practitioner Role Id
        business_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Business Status
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
        outputs:
          additionalProperties: true
          type: object
          title: Outputs
      additionalProperties: false
      type: object
      title: TaskActionRequest
    TaskFocusType:
      type: string
      enum:
      - order-item
      - specimen
      - observation
      - report
      title: TaskFocusType
    TaskSummary:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        group_identifier:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Identifier
        based_on_order_item_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Based On Order Item Id
        focus_type:
          $ref: '#/components/schemas/TaskFocusType'
        focus_id:
          type: string
          format: uuid
          title: Focus Id
        queue_code:
          type: string
          title: Queue Code
        status:
          type: string
          title: Status
        business_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Business Status
        priority:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority
        owner_user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner User Id
        authored_on:
          type: string
          format: date-time
          title: Authored On
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
      additionalProperties: false
      type: object
      required:
      - id
      - focus_type
      - focus_id
      - queue_code
      - status
      - authored_on
      title: TaskSummary
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer